Avoided an overflow in some extreme edge-case uses of WAVAudioFormat

+++ modules/juce_audio_formats/codecs/juce_WavAudioFormat.cpp
@@ -1258,7 +1258,7 @@ public:
         if (writeFailed)
             return false;
 
-        const size_t bytes = numChannels * (unsigned int) numSamples * bitsPerSample / 8;
+        const size_t bytes = numChannels * (size_t) numSamples * bitsPerSample / 8;
         tempBlock.ensureSize (bytes, false);
 
         switch (bitsPerSample)
